requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
alembic==1.7.7
    # via flask-migrate
click==8.1.3
    # via flask
colorama==0.4.4
    # via click
flask==2.1.2
    # via
    #   -r requirements.in
    #   flask-marshmallow
    #   flask-migrate
    #   flask-sqlalchemy
flask-marshmallow==0.14.0
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==1.1.2
    # via sqlalchemy
importlib-metadata==4.11.3
    # via
    #   alembic
    #   flask
importlib-resources==5.7.1
    # via alembic
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
marshmallow==3.15.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.28.0
    # via -r requirements.in
packaging==21.3
    # via marshmallow
pyparsing==3.0.9
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
six==1.16.0
    # via flask-marshmallow
sqlalchemy==1.4.36
    # via
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
werkzeug==2.1.2
    # via flask
zipp==3.8.0
    # via
    #   importlib-metadata
    #   importlib-resources
